+++ templates/parts/loop/loop.php
@@ -6,11 +6,16 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    1.0.0
- * @version  1.7.0
+ * @version  1.8.0
  */
 
 
 
+
+
+// Variables
+
+	$month = $last_month = '';
 
 
 ?>
@@ -20,6 +25,17 @@
 	<?php
 
 	while ( have_posts() ) : the_post();
+
+		$month = get_the_date( 'F Y' );
+
+		if ( $last_month !== $month ) {
+			echo '<div class="separator-month">';
+			echo '<a href="' . esc_url( get_month_link( date( 'Y', strtotime( $month ) ), date( 'm', strtotime( $month ) ) ) ) . '">';
+			echo '<span>' . implode( '</span> <span>', explode( ' ', $month ) ) . '</span>';
+			echo '</a>';
+			echo '</div>';
+			$last_month = $month;
+		}
 
 		get_template_part( 'templates/parts/content/content', 'simple' );